Move github.js and label.js to workflows
